requirements.txt
Django==2.1.2
django-jenkins==0.110.0
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.5
Pygments==2.2.0
pytz==2018.5